New gallery grid system
The grid system used on the main gallery has been reworked to both fit the  background and be responsive down to the point where the window height is the same as the width, where the window is supposed to display a message telling the user to rotate their device. There are also some slight adjustments to the images so that they cannot be dragged or selected.